backend/requirements.txt
Flask==2.3.2
torch==2.5.1
torchvision==0.20.1
Pillow==11.0.0
